Fix browser update msg on old browsers
